requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Vector DB
qdrant-client>=1.7.0

# GraphRAG
lightrag-hku>=1.4.0,<2.0.0

# Document Processing
python-multipart>=0.0.6

# Database (Metadata)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Frontend
streamlit>=1.30.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0

# Jupyter (Optional)
jupyter>=1.0.0
ipykernel>=6.29.0